requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alluka==0.1.0
async-timeout==4.0.2
attrs==21.4.0
Brotli==1.0.9
cchardet==2.1.7
cffi==1.15.0
charset-normalizer==2.0.12
ciso8601==2.2.0
colorama==0.4.4
colorlog==6.6.0
dnspython==2.2.1
frozenlist==1.3.0
hikari==2.0.0.dev107
hikari-tanjun==2.4.2a1
idna==3.3
motor==2.5.1
multidict==6.0.2
pycares==4.1.2
pycparser==2.21
pymongo==3.12.3
yarl==1.7.2
